refactor(data): Migration from `Double` to `BigDecimal` for financial values
Refactored the data layer and associated UI components to use `BigDecimal` instead of `Double` for representing financial amounts (balances, rates, and transaction values). This ensures better precision and prevents rounding errors common with floating-point arithmetic in financial calculations.

- **Database Schema Update:**
    - Updated `MizanDatabase` (version 8) to change the affinity of financial columns from `REAL` to `TEXT` across multiple tables, including `dashboard_summary`, `category_spending`, `transactions`, `accounts`, and `budgets`.
    - Affected fields include `totalBalance`, `monthlyExpenses`, `amount`, `budgetLimit`, and `rateToBase`.

- **MVI Store & Domain Models:**
    - Updated `AccountsStore.State` and `AccountItem` to use `BigDecimal` for `totalBalance` and `balance`.
    - Adjusted `AccountsLoaded` message to transport `BigDecimal` values.

- **Data Seeding & Repositories:**
    - Refactored `MockDataSeeder` to use `BigDecimal` literals and conversions when generating mock transactions, accounts, and currencies.
    - Cleaned up duplicate imports in `CurrencyRepositoryImpl`.

- **UI Kit Components:**
    - Updated `PremiumTotalBalanceCard` and `AccountRow` to accept `BigDecimal` for balance parameters.
    - Removed redundant `BigDecimal` wrapping inside composables as the values are now passed with the correct type.

+++ ui-kit/src/main/java/dev/esbi/mizan/ui/components/account/AccountRow.kt
@@ -26,7 +26,7 @@ import java.math.BigDecimal
 fun AccountRow(
     id: Long,
     name: String,
-    balance: Double,
+    balance: BigDecimal,
     currencyCode: String,
     modifier: Modifier = Modifier,
     onClick: () -> Unit = {}
@@ -54,7 +54,7 @@ fun AccountRow(
 
             BalanceAmount(
                 modifier = Modifier.padding(horizontal = 12.dp),
-                balance = BigDecimal(balance),
+                balance = balance,
                 currency = currencyCode,
                 typography = MizanTheme.typography.bodyLg
             )
